% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\alias{test_optimizer}
\alias{test_objective}
\title{Test functions}
\usage{
test_optimizer(
  objective = test_objective,
  initial = 1,
  ...,
  parameter = 1,
  value = objective(parameter),
  seconds = 0,
  warning_prob = 0,
  error_prob = 0,
  warning_msg = "warning",
  error_msg = "error",
  call. = TRUE
)

test_objective(
  x,
  value = x,
  warning_prob = 0,
  error_prob = 0,
  warning_msg = "warning",
  error_msg = "error",
  call. = TRUE
)
}
\arguments{
\item{objective}{[\code{function}]\cr
The objective function.}

\item{initial}{[\code{numeric()}]\cr
The initial parameter vector.}

\item{...}{[any]\cr
Optionally additional arguments to be passed to \code{objective}.}

\item{parameter}{[any]\cr
Defines the output parameter.}

\item{value}{[any]\cr
Defines the output function value.}

\item{seconds}{[\code{numeric(1)}]\cr
A delay in number of seconds.}

\item{warning_prob}{[\code{numeric(1)}]\cr
The probability for throwing a warning.}

\item{error_prob}{[\code{numeric(1)}]\cr
The probability for throwing an error.}

\item{warning_msg}{[\code{character(1)}]\cr
The warning message.}

\item{error_msg}{[\code{character(1)}]\cr
The error message.}

\item{call.}{[\code{logical(1)}]\cr
Passed to \code{\link[base]{warning}} or \code{\link[base]{stop}},
respectively.}

\item{x}{[any]\cr
Defines the input.}
}
\value{
A \code{list} with elements \code{parameter} and \code{value}.
}
\description{
These functions are useful for testing or debugging the behavior of objective
and optimization functions.

They can throw warnings or errors on purpose.
}
