% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_dictionary.R
\name{install_optimizer_packages}
\alias{install_optimizer_packages}
\title{Installing optimizer packages}
\usage{
install_optimizer_packages()
}
\value{
No return value.
}
\description{
This function installs packages that implement optimization algorithms.
}
