% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_number}
\alias{is_number}
\title{Check for number}
\usage{
is_number(x)
}
\arguments{
\item{x}{A (vector of) value(s).}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
This function checks if the input \code{x} is a (vector of) number(s), i.e.
a (vector of) positive integer value(s).
}
\examples{
\dontrun{
is_number(c(0, 1, 1.5))
}

}
\keyword{internal}
\keyword{utils}
