% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_portfolio.R
\name{optimize_portfolio}
\alias{optimize_portfolio}
\title{Portfolio Optimizer}
\usage{
optimize_portfolio(input_data, mthd = "mean-variance", target = 5)
}
\arguments{
\item{input_data}{Market data from which to make the portfolio.}

\item{mthd}{The method to be used in the optimization.}

\item{target}{The wanted return in percentages.}
}
\value{
A pie chart that visualizes the weights.
Also prints a report on the results
}
\description{
Optimize a portfolio using various methods
}
\examples{
temp1 <- optimize_portfolio(data)
}
