% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\docType{data}
\name{optimizer_nlm}
\alias{optimizer_nlm}
\alias{optimizer_optim}
\title{optim optimizer}
\usage{
optimizer_nlm

optimizer_optim
}
\description{
A short description...

Bla bla
}
\keyword{datasets}
