% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_objective}
\alias{test_objective}
\title{TODO}
\usage{
test_objective(x, warning_prob = 0, error_prob = 0)
}
\arguments{
\item{x}{TODO}

\item{warning_prob}{TODO}

\item{error_prob}{TODO}
}
\description{
TODO
}
