% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_optimizer}
\alias{test_optimizer}
\title{TODO}
\usage{
test_optimizer(
  f,
  p,
  ...,
  parameter = p,
  value = f(parameter),
  seconds = 0,
  warning_prob = 0,
  error_prob = 0
)
}
\arguments{
\item{f}{TODO}

\item{p}{TODO}

\item{...}{TODO}

\item{parameter}{TODO}

\item{value}{TODO}

\item{seconds}{TODO}

\item{warning_prob}{TODO}

\item{error_prob}{TODO}
}
\description{
TODO
}
