% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timed}
\alias{timed}
\title{Interruption of long evaluations}
\usage{
timed(expr, secs)
}
\arguments{
\item{expr}{An R expression to evaluate.}

\item{secs}{The number of seconds.}
}
\value{
Either the value of \code{expr} or \code{NULL} if the evaluation time
exceeded \code{secs} seconds.
}
\description{
This function evaluates \code{expr} and interrupts the evaluation after
\code{secs} seconds.
}
\details{
This function is a wrapper for \code{\link[R.utils]{withTimeout}}.
}
\examples{
foo <- function(x) { Sys.sleep(x); return(x) }
timed(foo(0.5), 1)
timed(foo(1.5), 1)

}
\keyword{internal}
\keyword{utils}
