% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_silent}
\alias{try_silent}
\title{Try an expression silently}
\usage{
try_silent(expr)
}
\arguments{
\item{expr}{An R expression to try.}
}
\value{
Either the value of \code{expr} or in case of a failure an object of class
\code{fail}, which contains the error message.
}
\description{
This function tries to execute \code{expr} and returns a string with the
error message if the execution failed.
}
\details{
This function is a wrapper for \code{\link[base]{try}}.
}
\examples{
try_silent(1 + 1)
try_silent(1 + "1")

}
\keyword{internal}
\keyword{utils}
