% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeR-package.R
\docType{package}
\name{optimizeR-package}
\alias{optimizeR}
\alias{optimizeR-package}
\title{optimizeR: Unified Framework for Numerical Optimizers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a unified framework for numerical optimizers in R, particularly for their inputs and outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/optimizeR/}
  \item \url{https://github.com/loelschlaeger/optimizeR/}
  \item Report bugs at \url{https://github.com/loelschlaeger/optimizeR/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

Other contributors:
\itemize{
  \item Marius Ötting \email{marius.oetting@uni-bielefeld.de} (\href{https://orcid.org/0000-0002-9373-0365}{ORCID}) [contributor]
}

}
\keyword{internal}
