% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{Optimizer}
\alias{Optimizer}
\title{Specify numerical optimizer}
\value{
An \code{Optimizer} object.
}
\description{
This function specifies the framework for a numerical optimizer.
}
\examples{
### define optimizer
optimizer_nelder_mead <- Optimizer$
  new(
    algorithm = pracma::nelder_mead  # optimization function
    tol = 1e-6                       # additional optimizer argument
  )$
  labels(
    objective = "fn",                # name of function input
    initial = "x0",                  # name of initial input
    value = "fmin",                  # name of value output
    parameter = "xmin"               # name of parameter output
  )

### define objective function
objective <- TestFunctions::TF_ackley

### optimize
optimizer_nelder_mead$apply(objective, initial = 1:2)

### compare to another optimizer
optimizer_nlm$apply(objective, initial = 1:2)
}
\seealso{
Wrappers for popular optimization algorithms in R are already available,
see for example \code{\link[=optimizer_nlm]{optimizer_nlm()}} for the \code{\link[stats]{nlm}} optimizer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Optimizer-new}{\code{Optimizer$new()}}
\item \href{#method-Optimizer-labels}{\code{Optimizer$labels()}}
\item \href{#method-Optimizer-validate}{\code{Optimizer$validate()}}
\item \href{#method-Optimizer-apply}{\code{Optimizer$apply()}}
\item \href{#method-Optimizer-print}{\code{Optimizer$print()}}
\item \href{#method-Optimizer-summary}{\code{Optimizer$summary()}}
\item \href{#method-Optimizer-clone}{\code{Optimizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$new(algorithm, name = NULL, direction = "min", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algorithm}}{}

\item{\code{name}}{}

\item{\code{direction}}{}

\item{\code{...}}{Additional arguments to be passed to the algorithm. Without
specifications, the default values are used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-labels"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-labels}{}}}
\subsection{Method \code{labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$labels(objective, initial, value, parameter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$validate(
  objective = test_objective,
  initial = 1:2,
  ...,
  seconds = 10
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-apply"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$apply(objective, initial, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-print"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$summary(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
