% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_dictionary.R
\docType{data}
\name{optimizer_dictionary}
\alias{optimizer_dictionary}
\title{Dictionary of optimizer functions}
\format{
An \code{R6} object of class \code{\link[oeli]{Dictionary}}.
}
\usage{
optimizer_dictionary
}
\description{
The \code{optimizer_dictionary} object is a dictionary of currently implemented
numerical optimizer functions.
}
\keyword{datasets}
