% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_objective}
\alias{test_objective}
\title{Test objective function}
\usage{
test_objective(
  x,
  value = x,
  warning_prob = 0,
  error_prob = 0,
  warning_msg = "warning",
  error_msg = "error",
  call. = TRUE
)
}
\arguments{
\item{x}{Any input.}

\item{value}{The return value, any object.}

\item{warning_prob}{The probability for throwing a warning.}

\item{error_prob}{The probability for throwing an error.}

\item{warning_msg}{The warning message.}

\item{error_msg}{The error message.}

\item{call.}{Passed to \code{\link[base]{warning}} or \code{\link[base]{stop}},
respectively.}
}
\value{
The argument \code{value}.
}
\description{
This function is useful for testing or debugging the behavior of objective
functions. It can throw a warning and / or an error on purpose.
}
