% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_optimizer}
\alias{test_optimizer}
\title{test optimizer}
\usage{
test_optimizer(
  f,
  p,
  ...,
  parameter = p,
  value = f(parameter),
  seconds = 0,
  warning_prob = 0,
  error_prob = 0
)
}
\description{
A short description...
}
